➕ Create `function.parser.js`
